+++ Source/WinCopies.IO.Shared/RegistryItemInfo.cs
@@ -232,7 +232,7 @@ namespace WinCopies.IO
         public RegistryItemInfo(RegistryKey registryKey) : base((registryKey ?? throw GetArgumentNullException(nameof(registryKey))).Name)
         {
 
-            string[] name = registryKey.Name.Split(System.IO.Path.PathSeparator);
+            string[] name = registryKey.Name.Split(WinCopies.IO.Path.PathSeparator);
 
             Name =
 
@@ -261,7 +261,7 @@ namespace WinCopies.IO
 
             ThrowIfNullEmptyOrWhiteSpace(path);
 
-            string[] name = path.Split(System.IO.Path.PathSeparator);
+            string[] name = path.Split(WinCopies.IO.Path.PathSeparator);
 
             Name =
 
@@ -385,7 +385,7 @@ namespace WinCopies.IO
 
                 case RegistryItemType.Key:
 
-                    string[] path = RegistryKey.Name.Split(System.IO.Path.PathSeparator);
+                    string[] path = RegistryKey.Name.Split(WinCopies.IO.Path.PathSeparator);
 
                     if (path.Length == 1)